+++ lib/builtin/settings_ui/settings_ui.dart
@@ -1,1 +1,2 @@
 export 'src/extension.dart';
+export 'src/settings_modal.dart';